Move Discord entities to a new namespace
