+++ src/Miunie.Core.XUnit.Tests/Providers/MiunieUserProviderTests.cs
@@ -1,4 +1,5 @@
-﻿using Miunie.Core.Providers;
+﻿using Miunie.Core.Entities.Discord;
+using Miunie.Core.Providers;
 using Miunie.Core.Storage;
 using Miunie.Core.XUnit.Tests.Data;
 using Moq;
