+++ src/Miunie.Core.XUnit.Tests/Providers/UserReputationProviderTests.cs
@@ -1,9 +1,10 @@
-using System;
-using Miunie.Core.Providers;
+using Miunie.Core.Entities.Discord;
 using Miunie.Core.Infrastructure;
-using Xunit;
-using Moq;
+using Miunie.Core.Providers;
 using Miunie.Core.XUnit.Tests.Data;
+using Moq;
+using System;
+using Xunit;
 
 namespace Miunie.Core.XUnit.Tests.Providers
 {
