+++ src/Miunie.Core.XUnit.Tests/Services/ProfileServiceTests.cs
@@ -1,4 +1,5 @@
 ﻿using Miunie.Core.Discord;
+using Miunie.Core.Entities.Discord;
 using Moq;
 using System.Threading.Tasks;
 using Xunit;
