+++ src/Miunie.Core.XUnit.Tests/Services/TimeServiceTests.cs
@@ -1,4 +1,5 @@
 ﻿using Miunie.Core.Discord;
+using Miunie.Core.Entities.Discord;
 using Miunie.Core.Infrastructure;
 using Miunie.Core.Providers;
 using Moq;
