+++ src/Miunie.Core/Discord/IDiscordConnection.cs
@@ -1,3 +1,4 @@
+using Miunie.Core.Entities.Discord;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
