+++ src/Miunie.Core/Discord/IDiscordMessages.cs
@@ -1,3 +1,4 @@
+using Miunie.Core.Entities.Discord;
 using System.Collections.Generic;
 using System.Threading.Tasks;
