+++ src/Miunie.Core/Entities/Discord/MiunieChannel.cs
@@ -1,4 +1,4 @@
-namespace Miunie.Core
+namespace Miunie.Core.Entities.Discord
 {
     public class MiunieChannel
     {
