+++ src/Miunie.Core/Entities/Discord/MiunieUser.cs
@@ -2,17 +2,11 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Miunie.Core
+namespace Miunie.Core.Entities.Discord
 {
     public class MiunieUser
     {
-        public Guid Id
-        {
-            get
-            {
-                return GenerateSeededGuid();
-            }
-        }
+        public Guid Id => GenerateSeededGuid();
 
         private Guid GenerateSeededGuid()
         {
@@ -26,20 +20,12 @@ namespace Miunie.Core
         public ulong GuildId { get; set; }
         public ulong UserId { get; set; }
         public Reputation Reputation { get; set; }
-        public List<ulong> NavCursor { get; set; }
         public DateTime JoinedAt { get; set; }
         public DateTime CreatedAt { get; set; }
         public bool IsBot { get; set; }
         public IEnumerable<MiunieRole> Roles { get; set; }
         public string AvatarUrl { get; set; }
         public TimeSpan? UtcTimeOffset { get; set; }
-
-        public MiunieUser()
-        {
-            NavCursor = new List<ulong>();
-        }
-
-
     }
 }
