--- src/Miunie.Core/Entities/LanguageResourceCollection.cs
@@ -1,9 +1,0 @@
-namespace Miunie.Core
-{
-    public class LanguageResourceCollection
-    {
-        public string Language;
-        public LangResource[] Resources;
-    }
-}
-
