+++ src/Miunie.Core/Providers/IMiunieUserProvider.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using Miunie.Core.Entities.Discord;
+using System.Collections.Generic;
 
 namespace Miunie.Core.Providers
 {
