+++ src/Miunie.Core/Providers/ITimeManipulationProvider.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Miunie.Core.Entities.Discord;
+using System;
 
 namespace Miunie.Core.Providers
 {
