+++ src/Miunie.Core/Providers/IUserReputationProvider.cs
@@ -1,3 +1,4 @@
+using Miunie.Core.Entities.Discord;
 using System.Collections.Generic;
 
 namespace Miunie.Core.Providers
