+++ src/Miunie.Core/Providers/MiunieUserProvider.cs
@@ -1,3 +1,4 @@
+using Miunie.Core.Entities.Discord;
 using Miunie.Core.Storage;
 using System.Collections.Generic;
