+++ src/Miunie.Core/Providers/UserReputationProvider.cs
@@ -1,8 +1,9 @@
+using Miunie.Core.Entities.Discord;
+using Miunie.Core.Infrastructure;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
-using Miunie.Core.Infrastructure;
 
 namespace Miunie.Core.Providers
 {
