+++ src/Miunie.Core/Services/RemoteRepositoryService.cs
@@ -1,5 +1,6 @@
 using Miunie.Core.Attributes;
 using Miunie.Core.Discord;
+using Miunie.Core.Entities.Discord;
 using System.Threading.Tasks;
 
 namespace Miunie.Core
