+++ src/Miunie.Core/Services/TimeService.cs
@@ -1,5 +1,6 @@
 ﻿using Miunie.Core.Attributes;
 using Miunie.Core.Discord;
+using Miunie.Core.Entities.Discord;
 using Miunie.Core.Infrastructure;
 using Miunie.Core.Providers;
 using System;
