+++ src/Miunie.Discord/Convertors/MiunieUserConvertor.cs
@@ -1,5 +1,5 @@
 using Discord.WebSocket;
-using Miunie.Core;
+using Miunie.Core.Entities.Discord;
 using Miunie.Core.Providers;
 using System;
 using System.Linq;
