+++ src/Miunie.Discord/Embeds/EmbedConstructor.cs
@@ -1,5 +1,6 @@
 ﻿using Discord;
 using Miunie.Core;
+using Miunie.Core.Entities.Discord;
 using Miunie.Core.Providers;
 using System;
 using System.Collections.Generic;
